seperate message file
